Introduce PostcodeLookUp
This object expects a postcode to be passed to it and returns
a tuple of `[longitude, latitude]`.

If it receives an invalid postcode or an empty field, it
raises `PostcodeLookUp.PostcodeNotFoundOrInvalid()`.

We'll later use this to calculate the coordinates of a community
energy company's address, and then save these to the database.

This will mean we can easily map them.