Bugfix: Make uuid readonly on edit, not create
tldr: if uuid is always readonly, it changes on first 'save'

Longer:

If UUID is readonly, when creating a new model, Django admin uses the
`default` method to create a new UUID, before the model is saved for the
first time.

However because it's readonly, the new UUID isn't POSTed on create, so
the new model is missing a UUID, and Django creates another one (again
using the `default` method)

In this scheme, you can edit the UUID on create, but never afterwards.